fix(web): restores toggle-module hotkeys, hotkeys in general
